--- packages/server/src/app/api/content/comments/comments.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CommentsController } from './comments.controller';
-
-describe('Comments Controller', () => {
-    let controller: CommentsController;
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            controllers: [CommentsController],
-        }).compile();
-
-        controller = module.get<CommentsController>(CommentsController);
-    });
-
-    it('should be defined', () => {
-        expect(controller).toBeDefined();
-    });
-});
