+++ packages/server/src/app/shared/content/comments.interface.ts
@@ -1,7 +1,32 @@
+import { PaginateResult } from 'mongoose';
+
+import { JwtPayload } from "@dragonfish/models/auth";
+import { ContentComment, CreateComment, EditComment } from "@dragonfish/models/comments";
+
 export interface IComments {
-    get()
+    /**
+     * Grabs the comments belonging to this content.
+     *
+     * @param contentId The content that these comments belong to
+     * @param pageNum The current page
+     */
+    get(contentId: string, pageNum: number): Promise<PaginateResult<ContentComment>>;
 
-    create()
+    /**
+     * Creates a new comment that belongs to some content.
+     *
+     * @param user A user's JWT payload
+     * @param contentId The content the comment belongs to
+     * @param commentInfo The comment's info
+     */
+    create(user: JwtPayload, contentId: string, commentInfo: CreateComment): Promise<ContentComment>;
 
-    edit()
+    /**
+     * Edits a comment belonging to a user, given its ID.
+     *
+     * @param user The owner of the comment
+     * @param commentId The comment's ID
+     * @param commentInfo The comment's new info
+     */
+    edit(user: JwtPayload, commentId: string, commentInfo: EditComment): Promise<ContentComment>;
 }
