+++ packages/server/src/app/shared/content/index.ts
@@ -1,3 +1,4 @@
 export { IContent } from './content.interface';
 export { ISections } from './sections.interface';
 export { ICollections } from './collections.interface';
+export { IComments } from './comments.interface';
